fix race in disconnect & synchronize snr

+++ internal/gateway/packets.go
@@ -110,7 +110,7 @@ type evtIdentity struct {
 type evtResume struct {
 	Token      string `json:"token"`
 	SessionID  string `json:"session_id"`
-	SequenceNr uint   `json:"seq"`
+	SequenceNr uint64 `json:"seq"`
 }
 
 type RequestGuildMembersPayload struct {
@@ -205,7 +205,7 @@ type helloPacket struct {
 type discordPacketJSON struct {
 	Op             opcode.OpCode `json:"op"`
 	Data           []byte        `json:"d"`
-	SequenceNumber uint          `json:"s"`
+	SequenceNumber uint64        `json:"s"`
 	EventName      string        `json:"t"`
 }
 
@@ -220,7 +220,7 @@ func (p *discordPacketJSON) CopyOverTo(packet *DiscordPacket) {
 type DiscordPacket struct {
 	Op             opcode.OpCode   `json:"op"`
 	Data           json.RawMessage `json:"d"`
-	SequenceNumber uint            `json:"s,omitempty"`
+	SequenceNumber uint64          `json:"s,omitempty"`
 	EventName      string          `json:"t,omitempty"`
 }
